Rename SizeClassificationAllocator to RangeAllocator
Also handle alignment properly
