+++ code/render/coregraphics/vk/vkgraphicsdevice.cc
@@ -1007,7 +1007,7 @@ CreateGraphicsDevice(const GraphicsDeviceCreateInfo& info)
         | CoreGraphics::BufferUsageFlag::TransferBufferDestination
         | CoreGraphics::BufferUsageFlag::ReadWriteBuffer;
     state.vertexBuffer = CoreGraphics::CreateBuffer(vboInfo);
-    state.vertexAllocator = Memory::SCAllocator(info.globalVertexBufferMemorySize, 0xFFFF);
+    state.vertexAllocator = Memory::RangeAllocator(info.globalVertexBufferMemorySize, 0xFFFF);
 
     CoreGraphics::BufferCreateInfo iboInfo;
     iboInfo.name = "Global Index Cache";
@@ -1019,7 +1019,7 @@ CreateGraphicsDevice(const GraphicsDeviceCreateInfo& info)
         | CoreGraphics::BufferUsageFlag::TransferBufferDestination
         | CoreGraphics::BufferUsageFlag::ReadWriteBuffer;
     state.indexBuffer = CoreGraphics::CreateBuffer(iboInfo);
-    state.indexAllocator = Memory::SCAllocator(info.globalIndexBufferMemorySize, 0xFFFF);
+    state.indexAllocator = Memory::RangeAllocator(info.globalIndexBufferMemorySize, 0xFFFF);
 
     CoreGraphics::BufferCreateInfo uploadInfo;
     uploadInfo.name = "Global Upload Buffer";
@@ -1677,7 +1677,7 @@ AllocateVertices(const SizeT numVertices, const SizeT vertexSize)
 {
     Threading::CriticalScope scope(&vertexAllocationMutex);
     const uint size = numVertices * vertexSize;
-    Memory::SCAlloc alloc = state.vertexAllocator.Alloc(size);
+    Memory::RangeAllocation alloc = state.vertexAllocator.Alloc(size);
     n_assert(alloc.offset != alloc.OOM);
     N_BUDGET_COUNTER_INCR(N_VERTEX_MEMORY, size);
     return VertexAlloc{ .size = size, .offset = alloc.offset, .node = alloc.node };
@@ -1690,7 +1690,7 @@ const VertexAlloc
 AllocateVertices(const SizeT bytes)
 {
     Threading::CriticalScope scope(&vertexAllocationMutex);
-    Memory::SCAlloc alloc = state.vertexAllocator.Alloc(bytes);
+    Memory::RangeAllocation alloc = state.vertexAllocator.Alloc(bytes);
     n_assert(alloc.offset != alloc.OOM);
     N_BUDGET_COUNTER_INCR(N_VERTEX_MEMORY, bytes);
     return VertexAlloc{ .size = (uint)bytes, .offset = alloc.offset, .node = alloc.node };
@@ -1703,7 +1703,7 @@ void
 DeallocateVertices(const VertexAlloc& alloc)
 {
     Threading::CriticalScope scope(&vertexAllocationMutex);
-    state.vertexAllocator.Dealloc(Memory::SCAlloc{.offset = alloc.offset, .node = alloc.node});
+    state.vertexAllocator.Dealloc(Memory::RangeAllocation{.offset = alloc.offset, .node = alloc.node});
     N_BUDGET_COUNTER_DECR(N_VERTEX_MEMORY, alloc.size);
 }
 
@@ -1724,7 +1724,7 @@ AllocateIndices(const SizeT numIndices, const IndexType::Code indexType)
 {
     Threading::CriticalScope scope(&vertexAllocationMutex);
     uint size = numIndices * IndexType::SizeOf(indexType);
-    Memory::SCAlloc alloc = state.indexAllocator.Alloc(size);
+    Memory::RangeAllocation alloc = state.indexAllocator.Alloc(size);
     n_assert(alloc.offset != alloc.OOM);
     N_BUDGET_COUNTER_INCR(N_INDEX_MEMORY, numIndices * IndexType::SizeOf(indexType));
     return VertexAlloc{ .size = size, .offset = alloc.offset, .node = alloc.node };
@@ -1737,7 +1737,7 @@ const VertexAlloc
 AllocateIndices(const SizeT bytes)
 {
     Threading::CriticalScope scope(&vertexAllocationMutex);
-    Memory::SCAlloc alloc = state.indexAllocator.Alloc(bytes);
+    Memory::RangeAllocation alloc = state.indexAllocator.Alloc(bytes);
     n_assert(alloc.offset != alloc.OOM);
     N_BUDGET_COUNTER_INCR(N_INDEX_MEMORY, bytes);
     return VertexAlloc{ .size = (uint)bytes, .offset = alloc.offset, .node = alloc.node };
@@ -1750,7 +1750,7 @@ void
 DeallocateIndices(const VertexAlloc& alloc)
 {
     Threading::CriticalScope scope(&vertexAllocationMutex);
-    state.indexAllocator.Dealloc(Memory::SCAlloc{.offset = alloc.offset, .node = alloc.node});
+    state.indexAllocator.Dealloc(Memory::RangeAllocation{.offset = alloc.offset, .node = alloc.node});
     N_BUDGET_COUNTER_DECR(N_INDEX_MEMORY, alloc.size);
 }
