+++ tests/testfoundation/sizeclassificationallocatortest.cc
@@ -5,7 +5,7 @@
 #include "stdneb.h"
 #include "util/array.h"
 #include "sizeclassificationallocatortest.h"
-#include "memory/sizeclassificationallocator.h"
+#include "memory/rangeallocator.h"
 
 namespace Test
 {
@@ -21,11 +21,11 @@ void
 SizeClassificationAllocatorTest::Run()
 {
     using namespace Memory;
-    SCAllocator allocator(256, 4);
-    SCAlloc alloc1 = allocator.Alloc(64);
-    SCAlloc alloc2 = allocator.Alloc(64);
-    SCAlloc alloc3 = allocator.Alloc(64);
-    SCAlloc alloc4 = allocator.Alloc(64);
+    RangeAllocator allocator(256, 4);
+    RangeAllocation alloc1 = allocator.Alloc(64);
+    RangeAllocation alloc2 = allocator.Alloc(64);
+    RangeAllocation alloc3 = allocator.Alloc(64);
+    RangeAllocation alloc4 = allocator.Alloc(64);
     VERIFY(alloc1.offset == 0);
     VERIFY(alloc2.offset == 64);
     VERIFY(alloc3.offset == 128);
@@ -34,11 +34,11 @@ SizeClassificationAllocatorTest::Run()
     // Freeing up two consequtive blocks of 64 should allow for one free of 128
     allocator.Dealloc(alloc2);
     allocator.Dealloc(alloc3);
-    SCAlloc alloc5 = allocator.Alloc(128);
+    RangeAllocation alloc5 = allocator.Alloc(128);
     VERIFY(alloc2.offset == 64);
 
     // If we allocate now, it should be oom
-    SCAlloc alloc6 = allocator.Alloc(64);
+    RangeAllocation alloc6 = allocator.Alloc(64);
     VERIFY(alloc6.offset == alloc6.OOM);
 
     // Empty the allocator
